refactor: reorganiza consulta de BusinessDates
- Application: actualiza GetBusinessDateHandler para el nuevo flujo de consulta
- Application: actualiza IBusinessDateRepository para consultar BusinessDate
- Application: agrega errores específicos para BusinessDates
- Application: agrega BusinessDateErrors con BUSINESS_DATE_NOT_FOUND
- Infrastructure: actualiza BusinessDateRepository para el nuevo contrato
- API: actualiza BusinessDatesController para manejar el nuevo flujo de consulta

+++ src/Prestamax.Infrastructure/Tenant/BusinessDates/BusinessDateRepository.cs
@@ -12,7 +12,7 @@ public sealed class BusinessDateRepository(
     AppDbContext context) : IBusinessDateRepository
 {
     public async Task<BusinessDate?> GetByOrganizationIdAsync(
-        long organizationId,
+        int organizationId,
         CancellationToken cancellationToken)
     {
         return await context
